backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-dotenv==1.1.1
groq==0.31.0
scikit-learn==1.7.1
apscheduler==3.10.4
fpdf2==2.8.3
notion-client==2.3.0
pyaudio==0.2.14
websocket-client==1.8.0
requests==2.32.5
